FastAPI/fastapi_rest_udemy/02_schemas_pydantic/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
asyncpg
psycopg2
psycopg2-binary
alembic
python-decouple
asyncpg
email-validator
passlib[bcrypt]

